requirements.txt
qdrant_client==1.16.2
requests==2.32.5
grpcio==1.76.0
grpcio-tools==1.76.0
grpcio-reflection==1.76.0
